build: remove temporary files after exiting from generate_appimage.sh script
The generate_appimage.sh script was writing temporary file within the /tmp/enea_appimage
folder. This folder was not wiped after script execution. This was causing issues when
having subsequent runs of the script. The script now wipes the temporary folder out
(but keeps logs in case they need to be analyzed)